+++ platform/command/event.go
@@ -61,3 +61,45 @@ func UnmarshalEvent(data []byte, e *Event) error {
 	e.Payload = &payload
 	return nil
 }
+
+type RawEvent struct {
+	CommandUUID string
+	Time        time.Time
+	DeviceUDID  string
+	Payload     []byte
+}
+
+// NewRawEvent returns a RawEvent with the current time.
+func NewRawEvent(cmd *RawCommand) *RawEvent {
+	event := RawEvent{
+		CommandUUID: cmd.CommandUUID,
+		Time:        time.Now().UTC(),
+		DeviceUDID:  cmd.UDID,
+		Payload:     cmd.Raw,
+	}
+	return &event
+}
+
+// MarshalRawEvent serializes a RawEvent to a protocol buffer wire format.
+func MarshalRawEvent(e *RawEvent) ([]byte, error) {
+	return proto.Marshal(&commandproto.Event{
+		Id:           e.CommandUUID, // Id isn't used anywhere, so it's repurposed for CommandUUID
+		Time:         e.Time.UnixNano(),
+		DeviceUdid:   e.DeviceUDID,
+		PayloadBytes: e.Payload,
+	})
+}
+
+// UnmarshalRawEvent parses a protocol buffer representation of data into
+// the RawEvent.
+func UnmarshalRawEvent(data []byte, e *RawEvent) error {
+	var pb commandproto.Event
+	if err := proto.Unmarshal(data, &pb); err != nil {
+		return errors.Wrap(err, "unmarshal pb Event")
+	}
+	e.CommandUUID = pb.Id
+	e.Time = time.Unix(0, pb.Time).UTC()
+	e.DeviceUDID = pb.DeviceUdid
+	e.Payload = pb.PayloadBytes
+	return nil
+}