+++ platform/command/service.go
@@ -2,6 +2,7 @@
 package command
 
 import (
+	mdmsvc "github.com/micromdm/micromdm/mdm"
 	"github.com/micromdm/micromdm/mdm/mdm"
 	"github.com/micromdm/micromdm/platform/pubsub"
 	"golang.org/x/net/context"
@@ -9,15 +10,24 @@ import (
 
 type Service interface {
 	NewCommand(context.Context, *mdm.CommandRequest) (*mdm.CommandPayload, error)
+	NewRawCommand(context.Context, *RawCommand) error
+	ClearQueue(ctx context.Context, udid string) error
+}
+
+// Queue is an MDM Command Queue.
+type Queue interface {
+	Clear(context.Context, mdmsvc.CheckinEvent) error
 }
 
 type CommandService struct {
 	publisher pubsub.Publisher
+	queue     Queue
 }
 
-func New(pub pubsub.Publisher) (*CommandService, error) {
+func New(pub pubsub.Publisher, queue Queue) (*CommandService, error) {
 	svc := CommandService{
 		publisher: pub,
+		queue:     queue,
 	}
 	return &svc, nil
 }